Moved block and entity narration out of NarrationUtils
